Add a sitemap registry.
This is a base implementation of a registry pattern that we could use for registering all the individual sitemaps. Not sure if the signature for the `add_sitemap()` method is exactly what we would want, but something similar that would allow us to register all of the sitemaps in one place would be useful for setting up rewrites, building out the index sitemap, etc.