Refactor sitemap registration pattern.
This makes the Core_Sitemaps_Registry a property of the Core_Sitemaps class and moves the responsibility for registering sitemaps from each sitemap provider to the main Core_Sitemaps class. This allows each provider to expose the properties required for setting up and rendering sitemaps without needing to have any knowledge of the registry at all so we're not passing around as many objects between parts of the system.