Add Users sitemap
This adds a `Core_Sitemaps_Users` sitemap provider that provides the rewrite regex and query functionality to render user sitemap pages.

This also updates the `Core_Sitemaps_Renderer` so that the render function accepts a prepared url_list so each sitemap provider can query the appropriate object type and handle it's own implementation for building out the list of urls for a sitemap urlset. All existing providers have been updated to make use of the same rendering pattern.