Make maximum sitemap URL value filterable by sitemap type.
This adds the function, `core_sitemaps_get_max_urls()`, which can be used to get the value of the `CORE_SITEMAPS_MAX_URLS` constant and allows it to be filtered based on the type of sitemap. The new function is now used in each sitemap provider where previously we were directly referecing the constant, and each one passes in the slug of the current provider, which is passed to the filter.