Move sitemap pages from the registry to sitemap providers.
This changes the way sitemaps are registered so that we now only store sitemap providers to the registry and then only caculate the number of pages when the index is rendered, by moving the responsibility for providing sitemap entry data from the registry to the sitemap provider base class.

This ensures queries required for determining the max number of pages for each sitemap sub-type are only run when the sitemap is being rendered, rather than when sitemaps are registered. It also allows for us to pass sitemap entry data directly into the index render rather than relying on pulling sitemap URL information out of the registry, separating concerns.