Refactor assertXMLEqual() to work around a bug in DOMDocument::loadXml().
The PHP Manual states that DOMDocument::loadXml() "Returns TRUE on success or FALSE on failure.".  However, that's not 100% true.  If the XML is not-well-formed it usually returns false.  But if it is not-well-formed because a namespace prefix is used but not declared, it incorrectly returns true.

Luckily, in all cases of non-well-formed XML (that I've checked), libxml_get_last_error() returns the error.  So, this refactors Test_WP_Sitemaps_Renderer::loadXml() (which is used by Test_WP_Sitemaps_Renderer::assertXMLEqual()) to check that instead of the return value from DOMDocument::loadXml().